Reorganize, document
